+++ src/main/java/ru/vyarus/dropwizard/guice/module/context/info/impl/InstallerItemInfoImpl.java
@@ -2,8 +2,8 @@ package ru.vyarus.dropwizard.guice.module.context.info.impl;
 
 import com.google.common.collect.Sets;
 import ru.vyarus.dropwizard.guice.module.context.ConfigItem;
+import ru.vyarus.dropwizard.guice.module.context.ConfigScope;
 import ru.vyarus.dropwizard.guice.module.context.info.InstallerItemInfo;
-import ru.vyarus.dropwizard.guice.module.installer.scanner.ClasspathScanner;
 
 import java.util.Set;
 
@@ -32,6 +32,6 @@ public class InstallerItemInfoImpl extends ItemInfoImpl implements InstallerItem
 
     @Override
     public boolean isFromScan() {
-        return getRegisteredBy().contains(ClasspathScanner.class);
+        return getRegisteredBy().contains(ConfigScope.ClasspathScan.getType());
     }
 }