+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/debug/renderer/ContextTreeRendererTest.groovy
@@ -12,6 +12,7 @@ import ru.vyarus.dropwizard.guice.bundle.GuiceyBundleLookup
 import ru.vyarus.dropwizard.guice.diagnostic.support.bundle.FooBundle
 import ru.vyarus.dropwizard.guice.diagnostic.support.features.FooModule
 import ru.vyarus.dropwizard.guice.diagnostic.support.features.FooResource
+import ru.vyarus.dropwizard.guice.module.context.ConfigScope
 import ru.vyarus.dropwizard.guice.module.context.debug.DiagnosticBundle
 import ru.vyarus.dropwizard.guice.module.context.debug.report.tree.ContextTreeConfig
 import ru.vyarus.dropwizard.guice.module.context.debug.report.tree.ContextTreeRenderer
@@ -19,6 +20,7 @@ import ru.vyarus.dropwizard.guice.module.installer.CoreInstallersBundle
 import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBootstrap
 import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBundle
 import ru.vyarus.dropwizard.guice.module.installer.feature.LifeCycleInstaller
+import ru.vyarus.dropwizard.guice.module.installer.scanner.ClasspathScanner
 import ru.vyarus.dropwizard.guice.module.support.conf.GuiceyConfigurator
 import ru.vyarus.dropwizard.guice.support.util.GuiceRestrictedConfigBundle
 import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp
@@ -26,6 +28,9 @@ import spock.lang.Specification
 
 import javax.inject.Inject
 import javax.ws.rs.Path
+
+import static ru.vyarus.dropwizard.guice.module.context.ConfigScope.Configurator
+import static ru.vyarus.dropwizard.guice.module.context.ConfigScope.allExcept
 
 /**
  * @author Vyacheslav Rusakov
@@ -376,6 +381,20 @@ class ContextTreeRendererTest extends Specification {
     │
     ├── CLASSPATH SCAN
     │   └── installer  FooInstaller                 (r.v.d.g.d.s.features)
+    │
+    └── CONFIGURATORS
+        ├── -disable   DisabledExtension            (r.v.d.g.c.d.r.ContextTreeRendererTest)
+        └── -disable   DisabledModule               (r.v.d.g.c.d.r.ContextTreeRendererTest)
+"""
+    }
+
+
+    def "Check configurators exclusive render"() {
+        expect:
+        render(new ContextTreeConfig()
+                .hideScopes(allExcept(Configurator))) == """
+
+    APPLICATION
     │
     └── CONFIGURATORS
         ├── -disable   DisabledExtension            (r.v.d.g.c.d.r.ContextTreeRendererTest)