+++ src/test/groovy/ru/vyarus/dropwizard/guice/diagnostic/AutoScanModeWithBundleDiagnosticTest.groovy
@@ -8,6 +8,7 @@ import ru.vyarus.dropwizard.guice.diagnostic.support.features.FooModule
 import ru.vyarus.dropwizard.guice.diagnostic.support.features.FooResource
 import ru.vyarus.dropwizard.guice.module.GuiceSupportModule
 import ru.vyarus.dropwizard.guice.module.GuiceyConfigurationInfo
+import ru.vyarus.dropwizard.guice.module.context.ConfigScope
 import ru.vyarus.dropwizard.guice.module.context.info.InstallerItemInfo
 import ru.vyarus.dropwizard.guice.module.installer.CoreInstallersBundle
 import ru.vyarus.dropwizard.guice.module.installer.feature.LifeCycleInstaller
@@ -65,8 +66,8 @@ class AutoScanModeWithBundleDiagnosticTest extends BaseDiagnosticTest {
 
         and: "correct scopes"
         info.getActiveScopes() == [Application, ClasspathScanner, CoreInstallersBundle, FooBundle] as Set
-        info.getItemsByScope(Application) as Set == [CoreInstallersBundle, FooBundle, FooModule, GuiceSupportModule] as Set
-        info.getItemsByScope(ClasspathScanner) as Set == [FooInstaller, FooResource] as Set
+        info.getItemsByScope(ConfigScope.Application) as Set == [CoreInstallersBundle, FooBundle, FooModule, GuiceSupportModule] as Set
+        info.getItemsByScope(ConfigScope.ClasspathScan) as Set == [FooInstaller, FooResource] as Set
         info.getItemsByScope(FooBundle) as Set == [FooBundleInstaller, FooBundleResource, FooBundleModule, FooBundleRelativeBundle] as Set
 
         and: "lifecycle installer was disabled"