+++ src/test/groovy/ru/vyarus/dropwizard/guice/diagnostic/ManualModeWithBundleDiagnosticTest.groovy
@@ -8,6 +8,7 @@ import ru.vyarus.dropwizard.guice.diagnostic.support.features.FooModule
 import ru.vyarus.dropwizard.guice.diagnostic.support.features.FooResource
 import ru.vyarus.dropwizard.guice.module.GuiceSupportModule
 import ru.vyarus.dropwizard.guice.module.GuiceyConfigurationInfo
+import ru.vyarus.dropwizard.guice.module.context.ConfigScope
 import ru.vyarus.dropwizard.guice.module.context.info.InstallerItemInfo
 import ru.vyarus.dropwizard.guice.module.installer.CoreInstallersBundle
 import ru.vyarus.dropwizard.guice.module.installer.feature.ManagedInstaller
@@ -49,11 +50,11 @@ class ManualModeWithBundleDiagnosticTest extends BaseDiagnosticTest {
 
         and: "correct scopes"
         info.getActiveScopes() == [Application, FooBundle] as Set
-        info.getItemsByScope(Application) as Set == [FooModule, ResourceInstaller, FooResource, FooBundle, GuiceSupportModule] as Set
+        info.getItemsByScope(ConfigScope.Application) as Set == [FooModule, ResourceInstaller, FooResource, FooBundle, GuiceSupportModule] as Set
         info.getItemsByScope(FooBundle) as Set == [FooBundleInstaller, FooBundleResource, FooBundleModule, FooBundleRelativeBundle] as Set
 
         and: "foo installer was disabled"
-        !info.getItemsByScope(Application).contains(FooInstaller)
+        !info.getItemsByScope(ConfigScope.Application).contains(FooInstaller)
         InstallerItemInfo fi = info.data.getInfo(FooInstaller)
         !fi.enabled
         fi.disabledBy == [Application] as Set