+++ src/test/groovy/ru/vyarus/dropwizard/guice/diagnostic/ManualModeWithLookupDiagnosticTest.groovy
@@ -8,6 +8,7 @@ import ru.vyarus.dropwizard.guice.diagnostic.support.features.FooModule
 import ru.vyarus.dropwizard.guice.diagnostic.support.features.FooResource
 import ru.vyarus.dropwizard.guice.module.GuiceSupportModule
 import ru.vyarus.dropwizard.guice.module.GuiceyConfigurationInfo
+import ru.vyarus.dropwizard.guice.module.context.ConfigScope
 import ru.vyarus.dropwizard.guice.module.context.info.InstallerItemInfo
 import ru.vyarus.dropwizard.guice.module.installer.feature.jersey.JerseyFeatureInstaller
 import ru.vyarus.dropwizard.guice.module.installer.feature.jersey.ResourceInstaller
@@ -53,8 +54,8 @@ class ManualModeWithLookupDiagnosticTest extends BaseDiagnosticTest {
 
         and: "correct scopes"
         info.getActiveScopes() == [Application, GuiceRestrictedConfigBundle, HK2DebugBundle, GuiceyBundleLookup] as Set
-        info.getItemsByScope(Application) as Set == [ResourceInstaller, FooResource, FooModule, GuiceSupportModule] as Set
-        info.getItemsByScope(GuiceyBundleLookup) as Set == [GuiceRestrictedConfigBundle, HK2DebugBundle] as Set
+        info.getItemsByScope(ConfigScope.Application) as Set == [ResourceInstaller, FooResource, FooModule, GuiceSupportModule] as Set
+        info.getItemsByScope(ConfigScope.BundleLookup) as Set == [GuiceRestrictedConfigBundle, HK2DebugBundle] as Set
         info.getItemsByScope(GuiceRestrictedConfigBundle) as Set == [GuiceRestrictedConfigBundle.GRestrictModule] as Set
         info.getItemsByScope(HK2DebugBundle) as Set == [JerseyFeatureInstaller, HK2DebugFeature, HK2DebugBundle.HK2DebugModule] as Set